delete quiz relationship

+++ client/src/containers/Quizzes.vue
@@ -36,7 +36,10 @@ import {
   deleteQuizById,
   updateQuizById,
 } from "@models/quizzes.firebase";
-import { getQuestionsByQuizId } from "@models/questions.firebase";
+import {
+  getQuestionsByQuizId,
+  deleteQuestionById,
+} from "@models/questions.firebase";
 export default {
   components: {
     HeaderCustom,
@@ -108,6 +111,19 @@ export default {
     },
     onDeleteQuiz(quiz) {
       this.$store.dispatch("actLoadingAction", async () => {
+        // delete questions relationship (quiz)
+        let errorDelete = false;
+        const questions = await getQuestionsByQuizId(quiz._id);
+        for (let i = 0; i < questions.length; i += 1) {
+          const deleteItem = await deleteQuestionById(questions[i]._id);
+          if (!deleteItem) errorDelete = true;
+        }
+        // handle error
+        if (errorDelete)
+          return this.$toast.error(
+            this.$store.state.string.E_UNKNOWN_ERROR_DETECT
+          );
+
         // delete quizzes by id
         const deleteItem = await deleteQuizById(quiz._id);
         if (!deleteItem)
@@ -127,7 +143,7 @@ export default {
         const questions = await getQuestionsByQuizId(quiz._id);
         if (questions.length === 0)
           return this.$toast.error(
-            this.$store.state.string.E_QUIZZES_NOT_FOUND
+            this.$store.state.string.E_QUESTIONS_NOT_FOUND
           );
 
         const quiz_code = await this.generateCodeQuiz();
